#PHRAS-947 time 4h prod/LazaretController uses the manipulator new route to empty the quarantine DELETE /api/v2/quarantine/
